First start for unit testing

+++ z/math/fibonacci.cpp
@@ -6,14 +6,14 @@ namespace z
 {
 	namespace math
 	{
-		double fibonacci(double n) noexcept
+		long long fibonacci(int n) noexcept
 		{
 			double PHI_n  = pow(phi, n);
 			double Nphi_n = pow(-phi + 1, n);
 
 			double fib_n  = (PHI_n - Nphi_n) / sqrt5;
 
-			return fib_n;
+			return (long long)fib_n;
 		}
 	}
 }
